+++ test/OpenTelemetry.Configuration.Declarative.Tests/DeclarativeConfigurationEventSourceTests.cs
@@ -703,7 +703,7 @@ public sealed class DeclarativeConfigurationEventSourceTests
     private static ReadOnlyDictionary<string, string?> ReadConfiguration(string yaml)
     {
         using var factory = new DeclarativeYamlTestFileFactory();
-        return DeclarativeConfigurationReader.Read(new FilePath(factory.CreateYamlFile(yaml)));
+        return DeclarativeConfigurationReader.Read(new FilePath(factory.CreateYamlFile(yaml))).FlatKeys;
     }
 
     private sealed class TestEventSourceBuilder : IOpenTelemetryBuilder